Added claim conversion and worldedit perms
- Added functionality to convert claims from the old system to the new
system
- Added dynamic permission updating based on the region the player is
in, if they are builder or jrbuilder and in a region they are able to
build in then they will receive worldedit perms, similarly when they are
not in a region they are able to build in their perms will be removed
again. Additionally permissions will be removed on log-off.